Broke out collection info into new antibody.py Added a derived 'antibody_id' field as a unique key that is the combination of the required antibody_name and the optional but recommended antibody_product_no field - populate this field in the _update method  This info is same as used to be how the display_title was derived so ... Changed display_title to just return antibody_id field
