+++ src/encoded/schemas/antibody.json
@@ -16,16 +16,21 @@
     ],
     "type": "object",
     "required": [ "antibody_name", "award", "lab"],
-    "identifyingProperties": ["uuid","aliases","antibody_name"],
+    "identifyingProperties": ["uuid","aliases","antibody_id"],
     "properties": {
         "schema_version": {
             "default": "1"
         },
         "antibody_name": {
-            "title": "Enzyme Name",
+            "title": "Antibody Name",
             "description": "A short name of the antibody.",
+            "type": "string"
+        },
+        "antibody_id": {
+            "description": "DON'T SUBMIT, auto-generated, use for referencing antibodies in other sheets.",
             "type": "string",
-            "uniqueKey": "antibody:antibody_name"
+            "uniqueKey": true,
+            "exclude_from": ["submit4dn", "FFedit-create"]
         },
         "description": {
             "title": "Description",
